Prevent overflow when uploading a file greater than 4 GB, which would result in enternally sending zeroes once the end of the file was reached.
BUG=139815

Review URL: https://chromiumcodereview.appspot.com/10905004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@154146 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/base/upload_element.h
@@ -122,7 +122,7 @@ class NET_EXPORT UploadElement {
 
   // The byte offset from the beginning of the element data. Used to track
   // the current position when reading data.
-  size_t offset_;
+  uint64 offset_;
 
   // The stream of the element data, if this element is of TYPE_FILE.
   FileStream* file_stream_;
